Revert "[UPSTREAM] Refactor fetching groups (#242) (#217)"
This reverts commit d82af31f714eadb344e13686d40328a7cbca5702.
